requirements.txt
yfinance>=0.2.36
pdfplumber>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
rich>=13.7.0
typer>=0.9.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
anthropic>=0.40.0
questionary>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
